refactor(utils): modify utils directory
